Data maps and code files
This is a documentation of the use of machine learning methods and DNA methylation for the prediction of nine related diseases, including code and data graphs. The data graphs and code files are as follows: with normal individual data-ADASYN-CNN, with normal individual sample data, without normal sample data, without normal sample data-ADASYN and normal individual vs. patient data.